Address review comments
- Pass array name into abstract class constructor instead of overriding
  getter method
- Rename abstract classes to remove "AbstractDense" from the start
- Do not use TaskType to define array names for InferenceResults classes
