+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/inference/results/DenseEmbeddingBitResults.java
@@ -31,22 +31,17 @@ import java.util.List;
  * }
  * </pre>
  */
-public final class DenseEmbeddingBitResults extends AbstractDenseEmbeddingBitResults {
+public final class DenseEmbeddingBitResults extends EmbeddingBitResults {
     // This name is a holdover from before this class was renamed
     public static final String NAME = "text_embedding_service_bit_results";
     public static final String TEXT_EMBEDDING_BITS = "text_embedding_bits";
 
-    public DenseEmbeddingBitResults(List<AbstractDenseEmbeddingByteResults.Embedding> embeddings) {
-        super(embeddings);
+    public DenseEmbeddingBitResults(List<EmbeddingByteResults.Embedding> embeddings) {
+        super(embeddings, TEXT_EMBEDDING_BITS);
     }
 
     public DenseEmbeddingBitResults(StreamInput in) throws IOException {
-        super(in);
-    }
-
-    @Override
-    public String getArrayName() {
-        return TEXT_EMBEDDING_BITS;
+        super(in, TEXT_EMBEDDING_BITS);
     }
 
     @Override
